margin, padding fix
Until now paddings weren't respected.
The reason for this is that this information
is missing in the position information
in the children attribute of a canvas.
The new method to calculate the position of an
image placement uses the coords dict of an image canvas.
Urwid adds the entries of a coords dict of a child widget
to the parent widget.
Urwid adds paddings, etc. to the values of the coords dict.
So if a: (0, 0, somedata) is added to the coords dict of a child widgets canvas
the coords dict of the canvas of the root widget will contain the
position of this child widget.